Upserted
Changed Save from an add to an upsert. Also added ability to add new items to the list. Also also Update the list when save is complete.

TBH - it's a bit more chatty of an application than it needs to be right now, but we can refactor it if we get bored.